Renaming logging to logger
