+++ tower/logger/logging_thread.py
